Split out views/utils.py
